+++ components/application/PostAcceptanceForm.tsx
@@ -1,0 +1,3 @@
+export const PostAcceptanceForm: React.FC = () => {
+  return <>here is a form</>;
+};
